marmoraria-app/requirements.txt
flet
requests
certifi
urllib3
idna
charset-normalizer
python-dotenv
Pillow
fpdf2
defusedxml
fonttools